+++ docker/wsi-worker/svs2dz-functions.sh
@@ -4,13 +4,14 @@
 DID_ERROR=0
 JOB_IN_DIR=/data/job/in
 JOB_OUT_DIR=/data/job/out
+filename=$2
 
 print_help() {
   cat << EOL
 --- Command help
-Usage: svs2dz "KPMP_ID" "FILENAME_NO_EXTENSION" "PKG_FILE_ID" "SLIDE_TYPE" ["STAIN_TYPE"]
+Usage: svs2dz "KPMP_ID" "FILENAME_WITH_EXTENSION" "PKG_FILE_ID" "SLIDE_TYPE" ["STAIN_TYPE"]
 All arguments should be enclosed in double-quotes to escape spaces.
-Example: svs2dz "KPMP-Ex1" "KPMP-Ex1_TRI_1of1" "b9f7c729-8370-4f8d-9753-4a36e8ae57a4" "IF" "TRI"
+Example: svs2dz "KPMP-Ex1" "KPMP-Ex1_TRI_1of1.svs" "b9f7c729-8370-4f8d-9753-4a36e8ae57a4" "IF" "TRI"
 ... The above example creates symlinks at <ENV_WSE_LINK_FROM_DIR>/files_KPMP-Ex1/KPMP-Ex1_TRI_1of1, slide type Immunofluorescence, stain type Trichrome
 EOL
 }
@@ -45,7 +46,7 @@ validate_args() {
   fi
   
   if [[ -z $2 ]]; then
-    print_error "!!! arg 2 must be input file name (no extension) directly inside $JOB_IN_DIR"
+    print_error "!!! arg 2 must be input full file name directly inside $JOB_IN_DIR"
   fi
   
   if [[ -z $3 ]]; then
@@ -68,17 +69,12 @@ validate_args() {
 }
 
 call_vips() {
-  mkdir -p $ENV_LINK_SRC_DIR/files_$1
-  if [ "$4" == "EM" ] || [ "$4" == "em" ]; then
-    echo "--vips dzsave $JOB_IN_DIR/$2.jpg $ENV_LINK_SRC_DIR/files_$1/$2"
-    vips dzsave $JOB_IN_DIR/$2.jpg $ENV_LINK_SRC_DIR/files_$1/$2
-  else
-    echo "--- vips dzsave $JOB_IN_DIR/$2.svs $ENV_LINK_SRC_DIR/files_$1/$2"
-    vips dzsave $JOB_IN_DIR/$2.svs $ENV_LINK_SRC_DIR/files_$1/$2
-  fi
+  mkdir -p "$ENV_LINK_SRC_DIR/files_$1/${filename%.*}"
+  echo "--- vips dzsave $JOB_IN_DIR/$2 $ENV_LINK_SRC_DIR/files_$1/${filename%.*}"
+  vips dzsave "$JOB_IN_DIR/$2" "$ENV_LINK_SRC_DIR/files_$1/${filename%.*}"
   
   # Copy a consistently-well-sized DZ file out as our thumbnail
-  cp $ENV_LINK_SRC_DIR/files_$1/$2_files/8/0_0.jpeg $ENV_LINK_SRC_DIR/files_$1/tn_$2.jpeg
+  cp "$ENV_LINK_SRC_DIR/files_$1/${filename%.*}_files/8/0_0.jpeg" "$ENV_LINK_SRC_DIR/files_$1/tn_${filename%.*}.jpeg"
 }
 
 update_link_file() {
@@ -100,9 +96,9 @@ EOT
   
   cat <<EOT >> $JOB_OUT_DIR/svs2dz/link.sh
   if ! [ -L $ENV_LINK_DST_DIR/$3_files ]; then
-    ln -s $ENV_LINK_SRC_DIR/files_$1/$2_files $ENV_LINK_DST_DIR/$3_files
-    ln -s $ENV_LINK_SRC_DIR/files_$1/$2.dzi $ENV_LINK_DST_DIR/$3.dzi
-    ln -s $ENV_LINK_SRC_DIR/files_$1/tn_$2.jpeg $ENV_LINK_DST_DIR/tn_$3.jpeg
+    ln -s "$ENV_LINK_SRC_DIR/files_$1/${filename%.*}_files" "$ENV_LINK_DST_DIR/$3_files"
+    ln -s "$ENV_LINK_SRC_DIR/files_$1/${filename%.*}.dzi" "$ENV_LINK_DST_DIR/$3.dzi"
+    ln -s "$ENV_LINK_SRC_DIR/files_$1/tn_${filename%.*}.jpeg" "$ENV_LINK_DST_DIR/tn_$3.jpeg"
   fi
   
 EOT
@@ -116,7 +112,7 @@ run_link_file() {
 }
 
 extract_metadata() {
-  python3 /usr/sbin/extract_metadata.py $JOB_IN_DIR/$2.svs $JOB_OUT_DIR/metadata.json
+  python3 /usr/sbin/extract_metadata.py "$JOB_IN_DIR/$2" "$JOB_OUT_DIR/metadata.json"
 }
 
 generate_mongo_records() {